R08 almost done

+++ src/main/java/acme/forms/AuditorDashboard.java
@@ -28,8 +28,8 @@ public class AuditorDashboard extends AbstractForm {
 
 	private double				deviationPeriodInAuditRecords;
 
-	private int					minimumPeriodInAuditRecords;
+	private double				minimumPeriodInAuditRecords;
 
-	private int					maximumPeriodInAuditRecords;
+	private double				maximumPeriodInAuditRecords;
 
 }
